Fixed test breaking in openSUSE platform
By default, new users are created in group 100 in openSUSE (users), as
opposed to other OS' where the user is created in its own group. This
difference made this test to fail on openSUSE.

I'm unsure if the underlying issue may be on the `generate` function of
the `key` resource.