Fixed test breaking in openSUSE platform
By default, new users are created in group 100 in openSUSE (users), as
opposed to other OS' where the user is created in its own group. This
difference made this test to fail on openSUSE like this:

STDERR: /opt/chef/embedded/lib/ruby/gems/2.7.0/gems/mixlib-shellout-3.2.2/lib/mixlib/shellout.rb:221:in `getgrnam': can't find group for foo (ArgumentError)

I'm unsure if the underlying issue may be on the `generate` function of
the `key` resource... but this change works!